pyproject.toml
[project]
name = "medicalchatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.0.13",
    "flask>=3.1.1",
    "langchain>=0.3.26",
    "langchain-community>=0.3.26",
    "langchain-huggingface>=0.3.0",
    "openai>=1.91.0",
    "protobuf>=5.29.5",
    "pymupdf>=1.26.1",
    "pypdf>=5.6.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "sentence-transformers>=4.1.0",
    "streamlit>=1.46.0",
    "tiktoken>=0.9.0",
]
